:sparkles:(lld): fix what rebase broke
